chore: rename snapshot_with => with_snapshot and adjust docs (#39)
* chore: rename snapshow_with => with_snapshot

* adjust docs

* adjust docs

* update docs

* update docs

+++ benchmarks/src/format/common.rs
@@ -8,7 +8,7 @@ use rand::{
 #[derive(openmetrics_client::encoder::EncodeLabelSet)]
 pub struct Labels {
     method: Method,
-    status: u32,
+    status: u16,
 }
 
 impl Labels {
